requirements.txt
altgraph==0.17
auto-py-to-exe==2.7.5
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.4.5.2
cffi==1.14.1
chardet==3.0.4
Eel==0.12.4
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
idna==2.9
numpy==1.18.5
pandas==1.0.4
pefile==2019.4.18
pycparser==2.20
PyInstaller==4.10
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
requests==2.23.0
six==1.15.0
urllib3==1.25.9
whichcraft==0.6.1
xlwt==1.3.0
zope.event==4.4
zope.interface==5.1.0
PyYAML==6.0